+++ CHANGELOG.md
@@ -9,6 +9,7 @@
 - Use the up and down arrow keys to scroll code vertically.
 - Show the type of change in the overlapping code area's title.
 - When new and deleted files are empty, show `<contents empty>` in code area.
+- Use ▓ characters to demarcate diff sections when showing in code area.
 
 ### Changed
 - Automatically scroll to the top of the code area after selecting a new file or revision.